+++ src/com/tisawesomeness/minecord/ReactListener.java
@@ -1,7 +1,5 @@
 package com.tisawesomeness.minecord;
 
-import com.tisawesomeness.minecord.ReactMenu.Emote;
-
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.MessageReaction;
 import net.dv8tion.jda.api.entities.User;
@@ -13,6 +11,7 @@ import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import java.util.HashMap;
 
 public class ReactListener extends ListenerAdapter {
+
     @Override
     public void onMessageReactionAdd(MessageReactionAddEvent e) {
         // Make sure message matches
